Finish flash in read rc

+++ User/Inc/flash.h
@@ -11,9 +11,9 @@
 #define FLASH_WRITE_SECTOR5(Buffer,Size)  FlashWrite(SECTOR5_ADDR,Buffer,Size)
 
 uint8_t FlashEraseSector5();
-void FlashRead(uint32_t address, uint16_t* buffer, uint32_t size);
-uint8_t FlashWrite(uint32_t address, uint16_t* buffer, uint32_t size);
-uint8_t FlashWriteWithoutErase(uint32_t address, uint16_t* buffer, uint32_t size);
+void FlashRead(uint32_t address, uint8_t* buffer, uint32_t size);
+uint8_t FlashWrite(uint32_t address, uint8_t* buffer, uint32_t size);
+uint8_t FlashWriteWithoutErase(uint32_t address, uint8_t* buffer, uint32_t size);
